refactor(www): bulk replace withDefaultSSR imports
