+++ apps/www/pages/2020.js
@@ -1,4 +1,4 @@
 import Page from '../components/Overview/pages/2020'
-import withDefaultSSR from '../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../lib/apollo/helpers'
 
 export default withDefaultSSR(Page)
