+++ apps/www/pages/2021.js
@@ -1,4 +1,4 @@
 import Page from '../components/Overview/pages/2021'
-import withDefaultSSR from '../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../lib/apollo/helpers'
 
 export default withDefaultSSR(Page)
