+++ apps/www/pages/503.js
@@ -12,7 +12,7 @@ import {
 import { SPACE } from '../components/Frame/PureFooter'
 
 import { PUBLIC_BASE_URL, CDN_FRONTEND_BASE_URL } from '../lib/constants'
-import withDefaultSSR from '../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../lib/apollo/helpers'
 
 const styles = {
   container: css({
