+++ apps/www/pages/aktionariat.js
@@ -8,7 +8,7 @@ import Table from '../components/Shareholder/Table'
 import Sunburst, { radius } from '../components/Shareholder/Sunburst'
 
 import { CDN_FRONTEND_BASE_URL } from '../lib/constants'
-import withDefaultSSR from '../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../lib/apollo/helpers'
 
 const ShareholderPage = ({ t }) => {
   const meta = {
