+++ apps/www/pages/benachrichtigungen/index.js
@@ -7,7 +7,7 @@ import SignIn from '../../components/Auth/SignIn'
 import { Interaction } from '@project-r/styleguide'
 
 import { CDN_FRONTEND_BASE_URL } from '../../lib/constants'
-import withDefaultSSR from '../../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../../lib/apollo/helpers'
 
 const NotificationsPage = ({ t, me }) => {
   const meta = {
