+++ apps/www/pages/einrichten.js
@@ -1,6 +1,6 @@
 import Onboarding from '../components/Onboarding/Page'
 
 import { enforceMembership } from '../components/Auth/withMembership'
-import withDefaultSSR from '../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../lib/apollo/helpers'
 
 export default withDefaultSSR(enforceMembership()(Onboarding))
