+++ apps/www/pages/konto/index.js
@@ -8,7 +8,7 @@ import { mediaQueries, A, Interaction } from '@project-r/styleguide'
 import Frame from '../../components/Frame'
 import Merci from '../../components/Pledge/Merci'
 import withT from '../../lib/withT'
-import withDefaultSSR from '../../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../../lib/apollo/helpers'
 import AccountTabs from '../../components/Account/AccountTabs'
 import AccountSection from '../../components/Account/AccountSection'
 import Memberships from '../../components/Account/Memberships'
