+++ apps/www/pages/markdown.js
@@ -21,7 +21,7 @@ import {
 } from '@project-r/styleguide'
 
 import { PUBLIC_BASE_URL, CDN_FRONTEND_BASE_URL } from '../lib/constants'
-import withDefaultSSR from '../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../lib/apollo/helpers'
 
 const { P, A, Emphasis, Cursive, StrikeThrough } = Editorial
