+++ apps/www/pages/mitteilung.js
@@ -34,7 +34,7 @@ import {
   ColorHtmlBodyColors,
 } from '@project-r/styleguide'
 import Link from 'next/link'
-import withDefaultSSR from '../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../lib/apollo/helpers'
 
 const styles = {
   bar: css({
