+++ apps/www/pages/rubriken.js
@@ -6,7 +6,7 @@ import withT from '../lib/withT'
 import { CDN_FRONTEND_BASE_URL } from '../lib/constants'
 
 import { Center } from '@project-r/styleguide'
-import withDefaultSSR from '../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../lib/apollo/helpers'
 
 const FormatsPage = ({ t }) => {
   const meta = {
