+++ apps/www/pages/suche.js
@@ -6,7 +6,7 @@ import { enforceMembership } from '../components/Auth/withMembership'
 import withT from '../lib/withT'
 
 import { CDN_FRONTEND_BASE_URL } from '../lib/constants'
-import withDefaultSSR from '../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../lib/apollo/helpers'
 
 const SearchPage = ({ router, t }) => {
   const meta = {
