+++ apps/www/pages/updates/[[...slug]].js
@@ -1,7 +1,7 @@
 import Frame from '../../components/Frame'
 
 import List from '../../components/Updates/List'
-import withDefaultSSR from '../../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../../lib/apollo/helpers'
 
 const UpdatesPage = ({ serverContext }) => (
   <Frame raw>
