+++ apps/www/pages/veranstaltung/[slug].js
@@ -1,4 +1,4 @@
 import Page from '../../components/Events/Page'
-import withDefaultSSR from '../../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../../lib/apollo/helpers'
 
 export default withDefaultSSR(Page)
