+++ apps/www/pages/wahltindaer/index.js
@@ -26,7 +26,7 @@ import {
 } from '@project-r/styleguide'
 import { DiscussionIcon } from '@project-r/styleguide'
 import Link from 'next/link'
-import withDefaultSSR from '../../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../../lib/apollo/helpers'
 
 const query = gql`
   query getCardGroups {
