+++ apps/www/pages/wahltindaer/setup.js
@@ -8,7 +8,7 @@ import Meta from '../../components/Frame/Meta'
 
 import withT from '../../lib/withT'
 import { CDN_FRONTEND_BASE_URL } from '../../lib/constants'
-import withDefaultSSR from '../../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../../lib/apollo/helpers'
 
 const Page = ({ router, t }) => {
   const { token } = router.query
