+++ apps/www/pages/~/[slug].js
@@ -1,4 +1,4 @@
 import Profile from '../../components/Profile/Page'
-import withDefaultSSR from '../../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../../lib/apollo/helpers'
 
 export default withDefaultSSR(Profile)
